feat(sidebar): implement new sidebar component with navigation items and user section
