+++ frontend/src/shared/components/navigation/Sidebar/index.ts
@@ -1,0 +1,2 @@
+export { Sidebar } from './Sidebar';
+export type { SidebarProps, SidebarItem } from './sidebar.types';
